major improvements: text edition works in bi-directional mode (using DataBindings), with a special user interface that renders text in a normal fashion (FlowDocumentViewer), but accepts single local edits (at the TreeNode level) by mouse-clicking anywhere in the text (using a dynamic TextBox). When closing the window, the XUK is serialized into a separate document ({ORIGINAL_FILE}.out.xuk) to verify that the internal Urakawa SDK data model was indeed updated on the fly when editing the FlowDocument text.
Also fixed an important bug !! Removed lapsed DataModelEvent listeners explicitly, without even waiting for the garbage collector to kick-in (we could use the IDispose interface, but I think it's overkill here). Otherwise before this bug-fix, the PropertyChanged event was triggered many times for Run and TextBox widgets that were not even displayed anymore (the Binding was not aware of that).