the combo box selection works and updates are reflected in the data model.
however, there is still a problem in that when you change the type of a metadata item, validation is broken because the data template isn't changed as the item type changes.  this also applies to creating a new item, where you basically change an empty item into something of a given type.